+++ installers/uninstall.sh
@@ -6,7 +6,8 @@ set -e
 #                                                                                    #
 # Project 'pterodactyl-installer'                                                    #
 #                                                                                    #
-# Copyright (C) 2018 - 2022, Vilhelm Prytz, <vilhelm@prytznet.se>                    #
+# Copyright (C) 2018 - 2026, Vilhelm Prytz, <vilhelm@prytznet.se>                    #
+# Fork modifications Copyright (C) 2026, Ayanok0ji <https://github.com/Ayanok0ji>    #
 #                                                                                    #
 #   This program is free software: you can redistribute it and/or modify             #
 #   it under the terms of the GNU General Public License as published by             #
@@ -25,6 +26,8 @@ set -e
 #                                                                                    #
 # This script is not associated with the official Pterodactyl Project.               #
 # https://github.com/pterodactyl-installer/pterodactyl-installer                     #
+# Fork: https://github.com/Ayanok0ji/pterodactyl-installer                           #
+# Original project CCTO: Vilhelm Prytz & contributors                                #
 #                                                                                    #
 ######################################################################################
 
@@ -46,10 +49,10 @@ RM_WINGS="${RM_WINGS:-true}"
 rm_panel_files() {
   output "Removing panel files..."
   rm -rf /var/www/pterodactyl /usr/local/bin/composer
-  [ "$OS" != "centos" ] && unlink /etc/nginx/sites-enabled/pterodactyl.conf
-  [ "$OS" != "centos" ] && rm -f /etc/nginx/sites-available/pterodactyl.conf
-  [ "$OS" != "centos" ] && ln -s /etc/nginx/sites-available/default /etc/nginx/sites-enabled/default
-  [ "$OS" == "centos" ] && rm -f /etc/nginx/conf.d/pterodactyl.conf
+  [ "$OS" != "centos" ] && [ -L /etc/nginx/sites-enabled/pterodactyl.conf ] && unlink /etc/nginx/sites-enabled/pterodactyl.conf
+  [ "$OS" != "centos" ] && [ -f /etc/nginx/sites-available/pterodactyl.conf ] && rm -f /etc/nginx/sites-available/pterodactyl.conf
+  [ "$OS" != "centos" ] && [ ! -L /etc/nginx/sites-enabled/default ] && [ -f /etc/nginx/sites-available/default ] && ln -s /etc/nginx/sites-available/default /etc/nginx/sites-enabled/default
+  [ "$OS" == "centos" ] && [ -f /etc/nginx/conf.d/pterodactyl.conf ] && rm -f /etc/nginx/conf.d/pterodactyl.conf
   systemctl restart nginx
   success "Removed panel files."
 }
@@ -65,11 +68,12 @@ rm_docker_containers() {
 rm_wings_files() {
   output "Removing wings files..."
 
-  # stop and remove wings service
   systemctl disable --now wings
-  rm -rf /etc/systemd/system/wings.service
+  [ -f /etc/systemd/system/wings.service ] && rm -rf /etc/systemd/system/wings.service
 
-  rm -rf /etc/pterodactyl /usr/local/bin/wings /var/lib/pterodactyl
+  [ -d /etc/pterodactyl ] && rm -rf /etc/pterodactyl
+  [ -f /usr/local/bin/wings ] && rm -rf /usr/local/bin/wings
+  [ -d /var/lib/pterodactyl ] && rm -rf /var/lib/pterodactyl
   success "Removed wings files."
 }
 
@@ -98,7 +102,12 @@ rm_cron() {
 
 rm_database() {
   output "Removing database..."
-  valid_db=$(mysql -u root -e "SELECT schema_name FROM information_schema.schemata;" | grep -v -E -- 'schema_name|information_schema|performance_schema|mysql')
+  valid_db=$(mariadb -u root -e "SELECT schema_name FROM information_schema.schemata;" 2>/dev/null | grep -v -E -- 'schema_name|information_schema|performance_schema|mysql')
+  if [[ -z "$valid_db" ]]; then
+    warning "No valid databases found."
+    return
+  fi
+
   warning "Be careful! This database will be deleted!"
   if [[ "$valid_db" == *"panel"* ]]; then
     echo -n "* Database called panel has been detected. Is it the pterodactyl database? (y/N): "
@@ -111,19 +120,35 @@ rm_database() {
   else
     print_list "$valid_db"
   fi
-  while [ -z "$DATABASE" ] || [[ $valid_db != *"$database_input"* ]]; do
+
+  while [ -z "$DATABASE" ] || [[ "$valid_db" != *"$DATABASE"* ]]; do
     echo -n "* Choose the panel database (to skip don't input anything): "
     read -r database_input
     if [[ -n "$database_input" ]]; then
-      DATABASE="$database_input"
+      if [[ "$valid_db" == *"$database_input"* ]]; then
+        DATABASE="$database_input"
+      else
+        warning "Invalid database name. Try again."
+      fi
     else
       break
     fi
   done
-  [[ -n "$DATABASE" ]] && mysql -u root -e "DROP DATABASE $DATABASE;"
+
+  if [[ -n "$DATABASE" ]]; then
+    mariadb -u root -e "DROP DATABASE $DATABASE;" 2>/dev/null || warning "Failed to drop database $DATABASE."
+  else
+    output "No database selected, skipping removal."
+  fi
+
   # Exclude usernames User and root (Hope no one uses username User)
   output "Removing database user..."
-  valid_users=$(mysql -u root -e "SELECT user FROM mysql.user;" | grep -v -E -- 'user|root')
+  valid_users=$(mariadb -u root -e "SELECT user FROM mysql.user;" 2>/dev/null | grep -v -E -- 'user|root')
+  if [[ -z "$valid_users" ]]; then
+    warning "No valid database users found."
+    return
+  fi
+
   warning "Be careful! This user will be deleted!"
   if [[ "$valid_users" == *"pterodactyl"* ]]; then
     echo -n "* User called pterodactyl has been detected. Is it the pterodactyl user? (y/N): "
@@ -136,19 +161,31 @@ rm_database() {
   else
     print_list "$valid_users"
   fi
-  while [ -z "$DB_USER" ] || [[ $valid_users != *"$user_input"* ]]; do
+
+  while [ -z "$DB_USER" ] || [[ "$valid_users" != *"$DB_USER"* ]]; do
     echo -n "* Choose the panel user (to skip don't input anything): "
     read -r user_input
     if [[ -n "$user_input" ]]; then
-      DB_USER=$user_input
+      if [[ "$valid_users" == *"$user_input"* ]]; then
+        DB_USER=$user_input
+      else
+        warning "Invalid username. Try again."
+      fi
     else
       break
     fi
   done
-  [[ -n "$DB_USER" ]] && mysql -u root -e "DROP USER $DB_USER@'127.0.0.1';"
-  mysql -u root -e "FLUSH PRIVILEGES;"
-  success "Removed database and database user."
+
+  if [[ -n "$DB_USER" ]]; then
+    mariadb -u root -e "DROP USER '$DB_USER'@'127.0.0.1';" 2>/dev/null || warning "Failed to drop user $DB_USER."
+  else
+    output "No user selected, skipping removal."
+  fi
+
+  mariadb -u root -e "FLUSH PRIVILEGES;" 2>/dev/null
+  success "Removed database and database user (if selected)."
 }
+
 
 # --------------- Main functions --------------- #